Revert "Refactor endpoint package in Logic to request"
