+++ src/main/java/seedu/us/among/logic/Logic.java
@@ -6,8 +6,8 @@ import javafx.collections.ObservableList;
 import seedu.us.among.commons.core.GuiSettings;
 import seedu.us.among.logic.commands.CommandResult;
 import seedu.us.among.logic.commands.exceptions.CommandException;
+import seedu.us.among.logic.endpoint.exceptions.RequestException;
 import seedu.us.among.logic.parser.exceptions.ParseException;
-import seedu.us.among.logic.request.exceptions.RequestException;
 import seedu.us.among.model.Model;
 import seedu.us.among.model.ReadOnlyEndpointList;
 import seedu.us.among.model.endpoint.Endpoint;
