+++ src/main/java/seedu/us/among/logic/LogicManager.java
@@ -10,9 +10,9 @@ import seedu.us.among.commons.core.LogsCenter;
 import seedu.us.among.logic.commands.Command;
 import seedu.us.among.logic.commands.CommandResult;
 import seedu.us.among.logic.commands.exceptions.CommandException;
+import seedu.us.among.logic.endpoint.exceptions.RequestException;
 import seedu.us.among.logic.parser.ImposterParser;
 import seedu.us.among.logic.parser.exceptions.ParseException;
-import seedu.us.among.logic.request.exceptions.RequestException;
 import seedu.us.among.model.Model;
 import seedu.us.among.model.ReadOnlyEndpointList;
 import seedu.us.among.model.endpoint.Endpoint;
