+++ src/main/java/seedu/us/among/logic/commands/RunCommand.java
@@ -21,8 +21,8 @@ import com.fasterxml.jackson.core.JsonParseException;
 import seedu.us.among.commons.core.LogsCenter;
 import seedu.us.among.commons.util.StringUtil;
 import seedu.us.among.logic.commands.exceptions.CommandException;
-import seedu.us.among.logic.request.EndpointCaller;
-import seedu.us.among.logic.request.exceptions.RequestException;
+import seedu.us.among.logic.endpoint.EndpointCaller;
+import seedu.us.among.logic.endpoint.exceptions.RequestException;
 import seedu.us.among.model.Model;
 import seedu.us.among.model.endpoint.Endpoint;
 import seedu.us.among.model.endpoint.Response;
