+++ src/main/java/seedu/us/among/logic/endpoint/DeleteRequest.java
@@ -1,36 +1,36 @@
-package seedu.us.among.logic.request;
+package seedu.us.among.logic.endpoint;
 
 import java.io.IOException;
 
-import org.apache.http.client.methods.HttpOptions;
+import org.apache.http.client.methods.HttpDelete;
 import org.apache.http.client.methods.HttpUriRequest;
 
-import seedu.us.among.logic.request.exceptions.RequestException;
+import seedu.us.among.logic.endpoint.exceptions.RequestException;
 import seedu.us.among.model.endpoint.Endpoint;
 import seedu.us.among.model.endpoint.Response;
 
 /**
- * Contains the logic for sending options requests.
+ * Contains the logic for sending delete requests.
  */
-public class OptionsRequest extends Request {
+public class DeleteRequest extends Request {
 
     /**
-     * Constructor for OptionRequest.
+     * Constructor for DeleteRequest.
      *
      * @param endpoint endpoint to make API call on
      */
-    public OptionsRequest(Endpoint endpoint) {
+    public DeleteRequest(Endpoint endpoint) {
         super(endpoint);
     }
 
     /**
-     * Executes the API call with an options request.
+     * Executes the API call with a delete request.
      *
      * @return returns the response from the API call
      */
     @Override
     public Response send() throws IOException, RequestException {
-        HttpUriRequest request = new HttpOptions(super.getAddress());
+        HttpUriRequest request = new HttpDelete(super.getAddress());
         request = super.setHeaders(request, super.getHeaders());
         return super.execute(request);
     }
