+++ src/main/java/seedu/us/among/logic/endpoint/GetRequest.java
@@ -1,36 +1,36 @@
-package seedu.us.among.logic.request;
+package seedu.us.among.logic.endpoint;
 
 import java.io.IOException;
 
-import org.apache.http.client.methods.HttpHead;
+import org.apache.http.client.methods.HttpGet;
 import org.apache.http.client.methods.HttpUriRequest;
 
-import seedu.us.among.logic.request.exceptions.RequestException;
+import seedu.us.among.logic.endpoint.exceptions.RequestException;
 import seedu.us.among.model.endpoint.Endpoint;
 import seedu.us.among.model.endpoint.Response;
 
 /**
- * Contains the logic for sending head requests.
+ * Contains the logic for sending get requests.
  */
-public class HeadRequest extends Request {
+public class GetRequest extends Request {
 
     /**
-     * Constructor for HeadRequest.
+     * Constructor for GetRequest.
      *
      * @param endpoint endpoint to make API call on
      */
-    public HeadRequest(Endpoint endpoint) {
+    public GetRequest(Endpoint endpoint) {
         super(endpoint);
     }
 
     /**
-     * Executes the API call with a head request.
+     * Executes the API call with a get request.
      *
      * @return returns the response from the API call
      */
     @Override
     public Response send() throws IOException, RequestException {
-        HttpUriRequest request = new HttpHead(super.getAddress());
+        HttpUriRequest request = new HttpGet(super.getAddress());
         request = super.setHeaders(request, super.getHeaders());
         return super.execute(request);
     }
