+++ src/main/java/seedu/us/among/logic/endpoint/HeadRequest.java
@@ -1,40 +1,37 @@
-package seedu.us.among.logic.request;
+package seedu.us.among.logic.endpoint;
 
 import java.io.IOException;
 
-import org.apache.http.client.methods.HttpPost;
+import org.apache.http.client.methods.HttpHead;
 import org.apache.http.client.methods.HttpUriRequest;
 
-import seedu.us.among.logic.request.exceptions.RequestException;
+import seedu.us.among.logic.endpoint.exceptions.RequestException;
 import seedu.us.among.model.endpoint.Endpoint;
 import seedu.us.among.model.endpoint.Response;
 
 /**
- * Contains the logic for sending post requests.
+ * Contains the logic for sending head requests.
  */
-public class PostRequest extends Request {
+public class HeadRequest extends Request {
 
     /**
-     * Constructor for PostRequest.
+     * Constructor for HeadRequest.
      *
      * @param endpoint endpoint to make API call on
      */
-    public PostRequest(Endpoint endpoint) {
+    public HeadRequest(Endpoint endpoint) {
         super(endpoint);
     }
 
     /**
-     * Executes the API call with a post request.
+     * Executes the API call with a head request.
      *
      * @return returns the response from the API call
      */
     @Override
     public Response send() throws IOException, RequestException {
-        HttpUriRequest request = new HttpPost(super.getAddress());
-
+        HttpUriRequest request = new HttpHead(super.getAddress());
         request = super.setHeaders(request, super.getHeaders());
-        request = super.setData(request, super.getData());
-
         return super.execute(request);
     }
 }
