+++ src/main/java/seedu/us/among/logic/endpoint/PatchRequest.java
@@ -1,11 +1,11 @@
-package seedu.us.among.logic.request;
+package seedu.us.among.logic.endpoint;
 
 import java.io.IOException;
 
 import org.apache.http.client.methods.HttpPatch;
 import org.apache.http.client.methods.HttpUriRequest;
 
-import seedu.us.among.logic.request.exceptions.RequestException;
+import seedu.us.among.logic.endpoint.exceptions.RequestException;
 import seedu.us.among.model.endpoint.Endpoint;
 import seedu.us.among.model.endpoint.Response;
