+++ src/main/java/seedu/us/among/logic/endpoint/PostRequest.java
@@ -1,36 +1,36 @@
-package seedu.us.among.logic.request;
+package seedu.us.among.logic.endpoint;
 
 import java.io.IOException;
 
-import org.apache.http.client.methods.HttpPut;
+import org.apache.http.client.methods.HttpPost;
 import org.apache.http.client.methods.HttpUriRequest;
 
-import seedu.us.among.logic.request.exceptions.RequestException;
+import seedu.us.among.logic.endpoint.exceptions.RequestException;
 import seedu.us.among.model.endpoint.Endpoint;
 import seedu.us.among.model.endpoint.Response;
 
 /**
- * Contains the logic for sending put requests.
+ * Contains the logic for sending post requests.
  */
-public class PutRequest extends Request {
+public class PostRequest extends Request {
 
     /**
-     * Constructor for PutRequest.
+     * Constructor for PostRequest.
      *
      * @param endpoint endpoint to make API call on
      */
-    public PutRequest(Endpoint endpoint) {
+    public PostRequest(Endpoint endpoint) {
         super(endpoint);
     }
 
     /**
-     * Executes the API call with a put request.
+     * Executes the API call with a post request.
      *
      * @return returns the response from the API call
      */
     @Override
     public Response send() throws IOException, RequestException {
-        HttpUriRequest request = new HttpPut(super.getAddress());
+        HttpUriRequest request = new HttpPost(super.getAddress());
 
         request = super.setHeaders(request, super.getHeaders());
         request = super.setData(request, super.getData());
