+++ src/main/java/seedu/us/among/logic/endpoint/exceptions/RequestException.java
@@ -1,6 +1,6 @@
-package seedu.us.among.logic.request.exceptions;
+package seedu.us.among.logic.endpoint.exceptions;
 
-import seedu.us.among.logic.request.Request;
+import seedu.us.among.logic.endpoint.Request;
 
 /**
  * Represents an error which occurs during execution of a {@link Request}.
