+++ src/test/java/seedu/us/among/logic/LogicManagerTest.java
@@ -21,8 +21,8 @@ import seedu.us.among.logic.commands.AddCommand;
 import seedu.us.among.logic.commands.CommandResult;
 import seedu.us.among.logic.commands.ListCommand;
 import seedu.us.among.logic.commands.exceptions.CommandException;
+import seedu.us.among.logic.endpoint.exceptions.RequestException;
 import seedu.us.among.logic.parser.exceptions.ParseException;
-import seedu.us.among.logic.request.exceptions.RequestException;
 import seedu.us.among.model.Model;
 import seedu.us.among.model.ModelManager;
 import seedu.us.among.model.ReadOnlyEndpointList;
