+++ src/test/java/seedu/us/among/logic/commands/CommandTestUtil.java
@@ -16,7 +16,7 @@ import java.util.List;
 
 import seedu.us.among.commons.core.index.Index;
 import seedu.us.among.logic.commands.exceptions.CommandException;
-import seedu.us.among.logic.request.exceptions.RequestException;
+import seedu.us.among.logic.endpoint.exceptions.RequestException;
 import seedu.us.among.model.EndpointList;
 import seedu.us.among.model.Model;
 import seedu.us.among.model.endpoint.EndPointContainsKeywordsPredicate;
