+++ src/test/java/seedu/us/among/logic/endpoint/OptionsRequestTest.java
@@ -1,4 +1,4 @@
-package seedu.us.among.logic.request;
+package seedu.us.among.logic.endpoint;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
