+++ src/test/java/seedu/us/among/logic/endpoint/PostRequestTest.java
@@ -1,4 +1,4 @@
-package seedu.us.among.logic.request;
+package seedu.us.among.logic.endpoint;
 
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -15,18 +15,18 @@ import seedu.us.among.model.endpoint.Endpoint;
 import seedu.us.among.model.util.SampleDataUtil;
 import seedu.us.among.testutil.TypicalEndpoints;
 
-public class GetRequestTest {
+public class PostRequestTest {
 
-    private final Endpoint sampleValidEndpoint = TypicalEndpoints.GET;
-    private final Endpoint sampleHtmlResponseEndpoint = SampleDataUtil.getSampleHtmlResponseEndpoint("GET");
-    private final Endpoint sampleInvalidUrlEndpoint = SampleDataUtil.getSampleInvalidUrlEndpoint("GET");
-    private final GetRequest jsonRequest = new GetRequest(sampleValidEndpoint);
-    private final GetRequest htmlRequest = new GetRequest(sampleHtmlResponseEndpoint);
-    private final GetRequest invalidUrlRequest = new GetRequest(sampleInvalidUrlEndpoint);
+    private final Endpoint sampleValidEndpoint = TypicalEndpoints.POST;
+    private final Endpoint sampleHtmlResponseEndpoint = SampleDataUtil.getSampleHtmlResponseEndpoint("POST");
+    private final Endpoint sampleInvalidUrlEndpoint = SampleDataUtil.getSampleInvalidUrlEndpoint("POST");
+    private final PostRequest jsonRequest = new PostRequest(sampleValidEndpoint);
+    private final PostRequest htmlRequest = new PostRequest(sampleHtmlResponseEndpoint);
+    private final PostRequest invalidUrlRequest = new PostRequest(sampleInvalidUrlEndpoint);
 
     @Test
     public void constructor_null_throwsNullPointerException() {
-        assertThrows(NullPointerException.class, () -> new GetRequest(null));
+        assertThrows(NullPointerException.class, () -> new PostRequest(null));
     }
 
     @Test
